moving code around to make intention clearer.
