+++ src/crud/get-by-id.js
@@ -2,7 +2,7 @@
 var output = require('../output');
 var applyMaps = require('./step-maps');
 var _ = require('lodash');
-var addModel = require('../swagger/build-metadata/add-model');
+var addModel = require('../swagger/add-model');
 var config = require('nconf');
 
 module.exports = {
