+++ src/crud/update.js
@@ -1,9 +1,9 @@
 'use strict';
 var output = require('../output');
 var applyMaps = require('./step-maps');
-var ensureSchemaSet = require('./../swagger/build-metadata/ensure-schema-set');
+var ensureSchemaSet = require('./../metadata/ensure-schema-set');
 var getValidateFunction = require('./@shared/get-validate-function');
-var addModel = require('../swagger/build-metadata/add-model');
+var addModel = require('../swagger/add-model');
 var _ = require('lodash');
 var schemaName = 'update';
 var versionInfo = require('../version-info');
