+++ src/metadata/infer-names.js
